+++ src/coverlet.collector/Utilities/CoverletConstants.cs
@@ -27,5 +27,6 @@ namespace Coverlet.Collector.Utilities
     public const string DoesNotReturnAttributesElementName = "DoesNotReturnAttribute";
     public const string DeterministicReport = "DeterministicReport";
     public const string ExcludeAssembliesWithoutSources = "ExcludeAssembliesWithoutSources";
+    public const string ReportMerging = "ReportMerging";
   }
 }